norminette

+++ cub3D/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: cleblais <cleblais@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/22 14:30:42 by rthomas           #+#    #+#             */
-/*   Updated: 2023/04/19 13:35:50 by cleblais         ###   ########.fr       */
+/*   Updated: 2023/04/20 18:10:09 by cleblais         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s && s[i])
